Institute Management System/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-crontab==0.7.1
django-filter==21.1
django-on-heroku==1.1.2
django-storages==1.12.3
djangorestframework==3.12.4
dropbox==11.28.0
gunicorn==20.1.0
httpie==3.1.0
humanfriendly==10.0
idna==3.3
Markdown==3.3.4
multidict==6.0.2
Pillow==8.4.0
ply==3.11
psycopg2-binary==2.9.2
Pygments==2.12.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
stone==3.3.1
urllib3==1.26.7
whitenoise==5.3.0
